requirements.txt
selenium==3.141.0
pyyaml
requests==2.24.0